% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_fed_cepesp.R
\name{candidate_fed_cepesp}
\alias{candidate_fed_cepesp}
\title{Download data on the candidates' backgrounds in federal elections from Cepesp}
\usage{
candidate_fed_cepesp(position, year, ascii = FALSE, encoding = "Latin-1",
  export = FALSE)
}
\arguments{
\item{year}{Election year (\code{integer}). For this function, the following years are available: 1994, 1996, 1998,
2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014 and 2016.}

\item{ascii}{(\code{logical}). Should the text be transformed from Latin-1 to ASCII format?}

\item{encoding}{Data original encoding (defaults to 'windows-1252'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{export}{(\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?}
}
\value{
\code{voter_profile()} returns a \code{data.frame} with the following variables:

\itemize{
  \item PERIODO: Election year.
  \item UF: Units of the Federation's acronym in which occurred the election.
  \item MUNICIPIO: Municipality name.
  \item COD_MUNICIPIO_TSE: Municipal's Supreme Electoral Court code (number).
  \item NR_ZONA: Electoral zone's Supreme Electoral Court code (number).
  \item SEXO: Voters' sex.
  \item FAIXA_ETARIA: Voters' age group.
  \item GRAU_DE_ESCOLARIDADE: Voters' education degree.
  \item QTD_ELEITORES_NO_PERFIL: Absolute number of voters.
}
}
\description{
\code{candidate_fed_cepesp()} downloads and cleans data on candidates' backgrounds in federal elections from the API mantained by Cepesp/FVG.
The function returns a \code{data.frame} where each observation corresponds to a candidate.
}
\details{
If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 files in the current directory.
}
\examples{
\dontrun{
df <- candidate_fed_cepesp("presidente", 2002)
}
}
