% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate_fed_rda.R
\name{candidate_fed_rda}
\alias{candidate_fed_rda}
\title{Download data on the candidates' background in local elections}
\usage{
candidate_fed_rda(year)
}
\arguments{
\item{year}{Election year. For this function, onlye the years of 1996, 2000, 2004, 2008, 2012 and 2016
are available.}
}
\description{
Download data on the candidates' background in local elections
}
\examples{
\dontrun{
candidate_fed_rda(2000)
}
}
