% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility_code.R
\name{compatibility_code}
\alias{compatibility_code}
\title{Function for downloading compatibility codes with data from various public institutions, such as TSE, IBGE and Central Bank., by backup salve in platform OSF.}
\usage{
compatibility_code()
}
\value{
\code{compatibility_code()} returns a \code{data.frame}.
}
\description{
The \code{compatibility_code()} function is a wrapper compatibility codes with data from various public institutions, such as TSE, IBGE and Central Bank.
}
\examples{
\dontrun{
# Download data compatibility
code <- compatibility_code()
}
}
