% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections_cepesp.R
\name{elections_cepesp}
\alias{elections_cepesp}
\title{function elections data from cepesp API}
\usage{
elections_cepesp(year, type, position, data_table = FALSE)
}
\arguments{
\item{year}{From 1998 to 2018 for President, Governor, Senator, Federal Deputy, 
State Deputy or District Deputy , and  from 2000 to 2016 for Mayor or Councillor}

\item{type}{data type: candidate, vote}

\item{position}{President, Governor, Senator, Federal Deputy, State Deputy,
District Deputy, Mayor or Councillor}

\item{data_table}{return a data.table

 @return \code{elections_tse()} returns a \code{tibble}.}
}
\description{
function elections data from cepesp API
}
\examples{
\dontrun{
df <- elections_cepesp(2018, type = "candidate", "President")
}
}
