% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{elections_tse}
\alias{elections_tse}
\title{function elections}
\usage{
elections_tse(
  year,
  type,
  uf = "all",
  br_archive = FALSE,
  ascii = FALSE,
  encoding = "latin1",
  export = FALSE,
  temp = TRUE
)
}
\arguments{
\item{year}{Election year. For this function, only the years 1998, 2002, 2006, 2010, 2014, and 2018
are available.}

\item{type}{tipo de dados}

\item{uf}{Federation Unit acronym (\code{character vector}).}

\item{br_archive}{In the TSE's data repository, some results can be obtained for the whole country by loading a single
within a single file by setting this argument to \code{TRUE} (may not work in for some elections and, in 
other, it recoverns only electoral data for presidential elections, absent in other files).}

\item{ascii}{(\code{logical}). Should the text be transformed from Latin-1 to ASCII format?}

\item{encoding}{Data original encoding (defaults to 'Latin-1'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{export}{(\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?}

\item{temp}{(\code{logical}). If \code{TRUE}, keep the temporary compressed file for future use (recommended)}
}
\value{
\code{elections_tse()} returns a \code{data.frame} with the following variables:
}
\description{
\code{elections_tse()} downloads and aggregates data on the verification of federal elections in Brazil,
disaggregated by cities and electoral zone. The function returns a \code{data.frame} where each observation
corresponds to a city/zone.
}
\details{
If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 files in the current directory.
}
\note{
For the elections prior to 2002, some information can be incomplete. For the 2014 and 2018 elections, more variable are available.
}
\examples{
\dontrun{
df <- elections_tse(2002, type = "candidate")
}
}
