% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electionsBR.R
\docType{package}
\name{electionsBR-package}
\alias{electionsBR}
\alias{electionsBR-package}
\title{R functions to download and clean Brazilian electoral Data}
\description{
electionsBR provides a set of tools to easily pull and clean several different Brazilian electoral data:
election results; candidates' partisan affiliations, personal and professional backgrounds;
Brazilian parties' electoral performances; composition of electoral coalitions; among many
others.
}
\details{
To learn more about electionsBR, check the package vignette and documentation.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://electionsbr.com/}
  \item Report bugs at \url{https://github.com/silvadenisson/electionsBR/issues}
}

}
\author{
Beatriz Costa (UFMG/Brazil) [ctb], Fernando Meireles (UFMG/Brazil) [author],
Denisson Silva (UFMG/Brazil) [author]
}
