% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/electionsbr_rda.R
\name{electionsbr_rda}
\alias{electionsbr_rda}
\title{Download data on the candidates' background in local elections}
\usage{
electionsbr_rda(year, level = "fed", archive = "candidate")
}
\arguments{
\item{year}{Election year. For this function, onlye the years of 1996, 2000, 2004, 2008, 2012 and 2016
are available for local level. 1994, 1998, 2002, 2006, 2010, 2014, 2018 are available for fed level.}

\item{level}{Election level podem ser fed (default) or local.}

\item{archive}{archive corresponde a um da funcoes basicas do pacote, podem ser: candidate (default), vote_mun_zone,
legend, party_mun_zone, personal_finances, details_mun_zone and seats
atualizacao do arquivo a cada 1 ano}
}
\description{
Download data on the candidates' background in local elections
}
\examples{
\dontrun{
df <- electionsbr_rda(2018)
}
}
