% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{juntaDados}
\alias{juntaDados}
\title{Reads and rbinds multiple data.frames in the same directory}
\usage{
juntaDados(encoding)
}
\arguments{
\item{encoding}{Data original encoding (defaults to 'windows-1252'). This can be changed to avoid errors
when \code{ascii = TRUE}.}
}
\description{
Reads and rbinds multiple data.frames in the same directory
}

