% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parties_br}
\alias{parties_br}
\title{Returns a vector with the abbreviations of all Brazilian parties}
\usage{
parties_br()
}
\description{
The character vector includes only parties that ran in elections in 2016.
}
