% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_media.R
\name{social_media}
\alias{social_media}
\title{Download data on the candidates' links social media in federal elections}
\usage{
social_media(year, encoding = "latin1", temp = TRUE)
}
\arguments{
\item{year}{Election year (\code{integer}). For this function, only the years 2022
are available.}

\item{encoding}{Data original encoding (defaults to 'latin1'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{temp}{(\code{logical}). elections_rda}
}
\value{
\code{social_media()} returns a \code{tbl, data.frame}.
}
\description{
\code{social_media()} downloads data on the candidates' link social média 
federal elections in Brazil. The function returns a \code{data.frame} where each observation
corresponds to a candidates' links.
}
\details{
If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 files in the current directory.
}
\examples{
\dontrun{
df <- social_media(2022)
}
}
