% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/social_media_local.R
\name{social_media_local}
\alias{social_media_local}
\title{Download data on the candidates' links social media in federal elections}
\usage{
social_media_local(
  year,
  ascii = FALSE,
  encoding = "latin1",
  export = FALSE,
  temp = TRUE
)
}
\arguments{
\item{year}{Election year (\code{integer}). For this function, only the years 2020
are available.}

\item{ascii}{(\code{logical}). Should the text be transformed from latin1 to ASCII format?}

\item{encoding}{Data original encoding (defaults to 'latin1'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{export}{(\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?}

\item{temp}{(\code{logical}). elections_rda}
}
\value{
\code{social_media_local()} returns a \code{tbl, data.frame} with the following variables:

\itemize{
  \item DT_GERACAO: Generation date of the file (when the data was collected).
  \item HH_GERACAO: Generation time of the file (when the data was collected), Brasilia Time.
  \item ANO_ELEICAO: Election year.
  \item CD_TIPO_ELEICAO:
  \item NM_TIPO_ELEICAO:
  \item CD_ELEICAO:
  \item DS_ELEICAO: Description of the election.
  \item SQ_CANDIDATO: Candidate's sequence number generated internally by the electoral
  systems. It is not the candidate's campaign number.
  \item NR_ORDEM:
  \item DS_URL:
}
}
\description{
\code{social_media_local()} downloads data on the candidates' link social média 
federal elections in Brazil. The function returns a \code{data.frame} where each observation
corresponds to a candidates' links.
}
\details{
If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 files in the current directory.
}
\examples{
\dontrun{
df <- social_media_local(2020)
}
}
