% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_local_year}
\alias{test_local_year}
\title{Tests federal election year inputs}
\usage{
test_local_year(year)
}
\arguments{
\item{year}{A \code{numeric} vector of length == 1}
}
\description{
Tests federal election year inputs
}

