% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_mun_zone_local.R
\name{vote_mun_zone_local}
\alias{vote_mun_zone_local}
\title{Download data on the polls by candidate from local elections in Brazil}
\usage{
vote_mun_zone_local(year)
}
\arguments{
\item{year}{Election year. For this function, only the years 1996, 2000, 2004, 2008, and 2012
are available.}
}
\value{
\code{vote_mun_zone_local()} returns a \code{data.frame} with the following variables:

\itemize{
  \item DATA_GERACAO: Generation date of the file (when the data was collected).
  \item HORA_GERACAO: Generation time of the file (when the data was collected), Brasilia Time.
  \item ANO_ELEICAO: Election year.
  \item NUM_TURNO: Round number.
  \item DESCRICAO_ELEICAO: Description of the election.
  \item SIGLA_UF: Units of the Federation's acronym in which occurred the election.
  \item SIGLA_UE: Units of the Federation's acronym (In case of major election is the FU's
  acronym in which the candidate runs for (text) and in case of municipal election is the
  municipal's Supreme Electoral Court code (number)). Assume the special values BR, ZZ and
  VT to designate, respectively, Brazil, Overseas and Absentee Ballot.
  \item CODIGO_MUNICIPIO: Supreme Electoral code from the city where occurred the election.
  \item NOME_MUNICIPIO: Name of the city where occurred the election.
  \item NUMERO_ZONA: Zone number.
  \item CODIGO_CARGO: Code of the position that the candidate runs for.
  \item NUMERO_CANDIDATO: Candidate's number in the ballot box.
  \item SQ_CANDIDATO: Candidate's sequence number generated internally by the electoral
  \item NOME_CANDIDATO: Candidate's complete name.
  \item NOME_URNA_CANDIDATO: Candidate's ballot box name.
  \item DESCRICAO_CARGO: Description of the position that the candidate runs for.
  \item COD_SIT_CAND_TOT: Candidate's totalization status code in that election round.
  \item NUMERO_PARTIDO: Party number.
  \item SIGLA_PARTIDO: Party's acronym.
  \item NOME_PARTIDO: Party name.
  \item SEQUENCIAL_LEGENDA: Coalition's sequential number, generated internally by the electoral justice.
  \item NOME_COLIGACAO: COalition name.
  \item COMPOSICAO_LEGENDA: Coalition's composition.
  \item TOTAL_VOTOS: Total of votes.
 }
}
\description{
\code{vote_mun_zone_local()} downloads and aggregates data on the verification from local elections in Brazil,
disaggregated by electoral zone. The function returns a \code{data.frame} where each observation
corresponds to a city/zone.
}
\note{
For the elections prior to 2000, some information can be incomplete.
}
\examples{
\dontrun{
df <- vote_mun_zone_local(2000)
}
}

