% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_affiliation.R
\name{voter_affiliation}
\alias{voter_affiliation}
\title{Download data on voters' affiliation to political parties}
\usage{
voter_affiliation(party, uf, ascii = FALSE, encoding = "windows-1252")
}
\arguments{
\item{party}{\code{character} vector containing the acronym of the parties with an official record on TSE.
Spaces in the name of some parties (i.g. PC do B) must be preserved.}

\item{uf}{\code{character} vector containing the acronym of the states of the federation.}

\item{ascii}{(\code{logical}). Should the text be transformed from Latin-1 to ASCII format?}

\item{encoding}{Data original encoding (defaults to 'windows-1252'). This can be changed to avoid errors
when \code{ascii = TRUE}.}
}
\description{
\code{voter_affiliation()} downloads and cleans data on voters' affiliation to Brazilian political parties by state.
 The function returns a \code{data.frame} where each observation corresponds to a voter.
}
\note{
Data on party affiliation is uptadet on a daily basis, therefore it may vary depending on the day it is collected.
}
\examples{
\dontrun{
df <- voter_affiliation("PT", "DF")

df <- voter_affiliation(c("PT", "PC do B"), "DF")

df <- voter_affiliation(c("PT", "PC do B"), c("DF", "MG", "AL"))
}
}

