% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voter_profile_place.R
\name{voter_profile_place}
\alias{voter_profile_place}
\title{Download data on the voters' profile by section}
\usage{
voter_profile_place(
  year,
  ascii = FALSE,
  encoding = "windows-1252",
  export = FALSE,
  temp = TRUE
)
}
\arguments{
\item{year}{Election year (\code{integer}). For this function, the following years are available: 1994, 1996, 1998,
2000, 2002, 2004, 2006, 2008, 2010, 2012, 2014, 2016, 2018 and 2020.}

\item{ascii}{(\code{logical}). Should the text be transformed from Latin-1 to ASCII format?}

\item{encoding}{Data original encoding (defaults to 'windows-1252'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{export}{(\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?}

\item{temp}{(\code{logical}). If \code{TRUE} keep temporary compressed file}
}
\value{
\code{voter_profile()} returns a \code{data.frame} with the following variables:

\itemize{
  \item  DT_GERACAO
  \item  HH_GERACAO
  \item  AA_ELEICAO    
  \item  DT_ELEICAO            
  \item  DS_ELEICAO
  \item  SG_UF              
  \item  CD_MUNICIPIO      
  \item  NM_MUNICIPIO         
  \item  NR_ZONA        
  \item  NR_SECAO             
  \item  CD_TIPO_SECAO_AGREGADA
  \item  DS_TIPO_SECAO_AGREGADA
  \item  NR_LOCAL_VOTACAO 
  \item  NM_LOCAL_VOTACAO   
  \item  CD_TIPO_LOCAL    
  \item  DS_TIPO_LOCAL        
  \item  DS_ENDERECO    
  \item  NM_BAIRRO        
  \item  NR_CEP             
  \item  NR_TELEFONE_LOCAL   
  \item  NR_LATITUDE
  \item  NR_LONGITUDE          
  \item  CD_SITU_LOCAL_VOTACAO
  \item  DS_SITU_LOCAL_VOTACAO 
  \item  CD_SITU_ZONA        
  \item  DS_SITU_ZONA       
  \item  CD_SITU_SECAO    
  \item  DS_SITU_SECAO      
  \item  CD_SITU_LOCALIDADE
  \item  DS_SITU_LOCALIDADE   
  \item  QT_ELEITOR
  \item  QT_ELEITOR_ELEICAO"   
  
}
}
\description{
\code{voter_profile()} downloads and cleans data on the voters' profile aggregated by state, city and electoral zone.
The function returns a \code{data.frame} where each observation corresponds to a voter profile type.
}
\details{
If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 files in the current directory.
}
\examples{
\dontrun{
df <- voter_profile_place(2016)
}
}
