% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elections.R
\name{elections_tse}
\alias{elections_tse}
\title{function elections}
\usage{
elections_tse(
  year,
  type,
  uf = "all",
  br_archive = FALSE,
  ascii = FALSE,
  encoding = "latin1",
  export = FALSE,
  temp = TRUE,
  data_table = FALSE,
  readme_pdf = FALSE
)
}
\arguments{
\item{year}{Election year. From 1998 to 2022.
are available.}

\item{type}{tipo de dados 

\itemize{
  \item candidate: downloads the data on the candidates'. 
  Where each observation corresponds to a candidate.
  \item vote_mun_zone: downloads data on the verification, disaggregated by cities and electoral zone.
   Where each observation corresponds to a city/zone.
  \item details_mun_zone: downloads THE data on disaggregated by town and electoral zone.
  Where each observation corresponds to a town/zone.
  \item legends: downloads the data on the party denomination (coalitions or parties), disaggregated bi cities. 
  Where each observation corresponds to a city.
  \item party_mun_zone: downloads the data on the polls by parties, disaggregated by cities and electoral zones. 
  Where each observation corresponds to a city/zone.
  \item personal_finances: download the data on personal financial disclosures. 
  Where each observation corresponds to a candidate's property.
  \item seats: download data on the number of seats under dispute in elections
  \item vote_section: download data on candidate electoral results in elections in Brazil by electoral section
  \item vote_profile_by_section: download data on the voters' profile by vote section
  \item vote_profile: download data on the voters' profile
  \item social_media: download data on the candidates' links social media in federal elections
}}

\item{uf}{Federation Unit acronym (\code{character vector}).}

\item{br_archive}{In the TSE's data repository, some results can be obtained for the whole country by loading a single
within a single file by setting this argument to \code{TRUE} (may not work in for some elections and, in 
other, it recoverns only electoral data for presidential elections, absent in other files).}

\item{ascii}{(\code{logical}). Should the text be transformed from Latin-1 to ASCII format?}

\item{encoding}{Data original encoding (defaults to 'Latin-1'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{export}{(\code{logical}). Should the downloaded data be saved in .dta and .sav in the current directory?}

\item{temp}{(\code{logical}). If \code{TRUE}, keep the temporary compressed file for future use (recommended)}

\item{data_table}{for data.table}

\item{readme_pdf}{is original readme}
}
\value{
\code{elections_tse()} returns a \code{data.frame} with the following variables:
}
\description{
\code{elections_tse()} data download from TSE repository
}
\details{
If export is set to \code{TRUE}, the downloaded data is saved as .dta and .sav
 files in the current directory.
}
\note{
For the elections prior to 2002, some information can be incomplete. For the 2014 and 2018 elections, more variable are available.
}
\examples{
\dontrun{
df <- elections_tse(2002, type = "candidate")
}
}
