% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_section_local.R
\name{vote_section_local}
\alias{vote_section_local}
\title{Download data on candidate electoral results in local elections in Brazil by electoral section}
\usage{
vote_section_local(year, uf = "AC", encoding = "latin1", temp = TRUE)
}
\arguments{
\item{year}{Election year. For this function, only the years 1996, 2000, 2004, 2008, 2012, 2016 and 2020
are available.}

\item{uf}{Federation Unit acronym (\code{character vector}). Defaults to \code{'AC'} (Acre).}

\item{encoding}{Data original encoding (defaults to 'Latin-1'). This can be changed to avoid errors
when \code{ascii = TRUE}.}

\item{temp}{(\code{logical}). If \code{TRUE}, keep the temporary compressed file for future use (recommended)}
}
\value{
\code{vote_section_local()} returns a \code{data.frame}.
}
\description{
\code{vote_section_local()} downloads and cleans data on the verification of local elections in Brazil,
disaggregated by electoral section. Different from other electionsBR's functions, results are only extract for individual states, one at a time. The function returns a \code{data.frame} where each observation
corresponds to an electoral section in a given Brazilian state.
}
\details{
If export is set to \code{TRUE}, the data is saved as .dta and .sav
 files in the working directory.
}
\note{
For the elections prior to 2002, some information can be incomplete.
}
\examples{
\dontrun{
df <- vote_section_local(2000)
}
}
